review: address comment
